create additional file for CLI
